backend/requirements.txt
# FastAPI and ASGI server
fastapi[all]>=0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
alembic==1.13.1

# Pydantic settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Async tasks
celery[redis]==5.3.6
redis==5.0.1

# Other utilities
httpx==0.26.0

# Vertex AI and Embeddings
google-cloud-aiplatform>=1.38.0
sentence-transformers>=2.2.2
torch>=2.0.0

# Web Scraping
beautifulsoup4>=4.12.2
lxml>=5.1.0

# Gemini AI
google-generativeai>=0.3.2

# Image Processing
Pillow>=10.1.0
numpy>=1.24.3
opencv-python>=4.8.1.78

# Stable Diffusion
replicate>=0.22.0

# Vectorization
pypotrace== 0.3.0
scour==0.38.2

# SAM2 (Segment Anything Model 2)
# Note: Install via pip install git+https://github.com/facebookresearch/segment-anything-2.git
# Or uncomment below:
# git+https://github.com/facebookresearch/segment-anything-2.git
torchvision==0.16.0

# SVG Processing
svgwrite==1.4.3

# Payments
stripe==7.9.0
